+++ app/controllers/developers_controller.rb
@@ -3,7 +3,7 @@ class DevelopersController < ApplicationController
   before_action :require_new_developer!, only: %i[new create]
 
   def index
-    @developers_count = SignificantFigure.new(Developer.visible.count).rounded
+    @developers_count = SignificantFigure.new(Developer.actively_looking_or_open.count).rounded
     @query = DeveloperQuery.new(permitted_attributes([:developers, :query]).merge(user: current_user))
     @meta = Developers::Meta.new(query: @query, count: @developers_count)
     Analytics::SearchQuery.create!(permitted_attributes([:developers, :query]))
@@ -71,7 +71,6 @@ class DevelopersController < ApplicationController
   def developer_params
     params.require(:developer).permit(
       :name,
-      :available_on,
       :hero,
       :bio,
       :website,
@@ -91,6 +90,6 @@ class DevelopersController < ApplicationController
       location_attributes: [:city, :state, :country],
       role_type_attributes: RoleType::TYPES,
       role_level_attributes: RoleLevel::TYPES
-    )
+    ).merge(user_initiated: true)
   end
 end