+++ app/models/developer.rb
@@ -1,9 +1,10 @@
 class Developer < ApplicationRecord
-  include Availability
   include Avatarable
   include Developers::HasOnlineProfiles
   include Developers::Notifications
+  include Developers::PublicChanges
   include Developers::RichText
+  include Developers::SearchScore
   include HasBadges
   include HasSpecialties
   include Hashid::Rails
@@ -68,10 +69,9 @@ class Developer < ApplicationRecord
   end
 
   scope :actively_looking_or_open, -> { where(search_status: [:actively_looking, :open, nil]) }
-  scope :available, -> { where(available_on: ..Time.current.to_date) }
-  scope :available_first, -> { where.not(available_on: nil).order(:available_on) }
   scope :featured, -> { where("featured_at >= ?", FEATURE_LENGTH.ago).order(featured_at: :desc) }
   scope :newest_first, -> { order(created_at: :desc) }
+  scope :by_score, -> { order(search_score: :desc, created_at: :asc) }
   scope :product_announcement_notifications, -> { where(product_announcement_notifications: true) }
   scope :profile_reminder_notifications, -> { where(profile_reminder_notifications: true) }
   scope :visible, -> { where.not(search_status: :invisible).or(where(search_status: nil)) }
@@ -108,7 +108,6 @@ class Developer < ApplicationRecord
       location.missing_fields? ||
       role_level.missing_fields? ||
       role_type.missing_fields? ||
-      available_on.blank? ||
       scheduling_link.blank?
   end